pat.pattern
Closed pattern filtering was erroneously excluding new patterns of one state.
Some conditions specific to MusMinr are toggled off, as they should not appear in PatMinr.
There seems to be a problem when calling closuretest for more general patterns, as the number of occurrences is required. Toggled off for the moment.